Replace update-notifier-common with our own logic
update-notifier-common is no longer available in Debian 8 (Jessie). As
such, we update the timestamp file ourselves when running `apt-get update`.

This might lead to more update runs than strictly necessary if `apt-get
update` is performed outside of a chef run. But at least, we are sure that
our own state is always up-to-date.