[GH-183] Control "ignore_failure" for "apt update" when adding repo
Define a new "ignore_apt_update_failure" attribute for "apt_repository"
which sets the "ignore_faillure" when "apt update" is being called.

If "ignore_apt_update_failure" is "false" and repository is not
installed, current provisioning with fail. Also, the sources list
file is removed, so the next "provision" will try to install the
repository again, and fail if its unsuccessful too.

Default value is "true" so it does not break existing code.